+++ tests/corpus.py
@@ -32,7 +32,7 @@ from sqlalchemy import (
 )
 from sqlalchemy.orm import DeclarativeBase, Session
 
-from conda_pupae.dist_repodata import (
+from conda_pupae.translate import (
     FileDistribution,
     pypi_to_conda_name,
     requires_to_conda,
